refactor : change the constant.java file position
