+++ src/main/java/baseball/Game.java
@@ -5,7 +5,7 @@ import utils.RandomUtil;
 
 import java.util.List;
 
-import static baseball.Constant.*;
+import static constant.Constant.*;
 
 public class Game {
