+++ src/main/java/baseball/Hint.java
@@ -2,7 +2,7 @@ package baseball;
 
 import java.util.List;
 
-import static baseball.Constant.*;
+import static constant.Constant.*;
 
 
 public class Hint {
