+++ src/main/java/utils/RandomUtil.java
@@ -5,7 +5,7 @@ import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
 
-import static baseball.Constant.*;
+import static constant.Constant.*;
 
 public class RandomUtil {
